# openapi/guardian.yaml
openapi: 3.1.0
servers:
  - url: https://api.guardianpot.internal
info:
  title: Guardian API
  version: 0.1.0
  description: Guardian v1 management and canonical telemetry API.
  license:
    name: Proprietary internal software
    identifier: LicenseRef-Guardian-Internal
security:
  - bearerAuth: []
paths:
  /v1/auth/bootstrap:
    post:
      operationId: bootstrapLocalOwner
      summary: Atomically exchange the one-time bootstrap token for the sole owner
      description: >-
        TLS-only. The CLI-created 32-byte token expires after 15 minutes and is
        consumed atomically. This response returns the TOTP provisioning URI
        and ten recovery codes exactly once and does not create a session.
      security:
        - bootstrapToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthBootstrapRequest'
      responses:
        '201':
          description: The pending-MFA owner was created.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthBootstrapResult'
        '401':
          description: The bootstrap exchange was generically denied.
        '426':
          description: TLS is required.
  /v1/auth/login:
    post:
      operationId: loginLocalOwner
      summary: Establish a server-side owner session with password and MFA
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthSessionCredentialsResponse'
        '401':
          description: Authentication was generically denied.
        '429':
          description: Persistent account or source throttling denied the attempt.
        '426':
          description: TLS is required.
  /v1/auth/session:
    get:
      operationId: getLocalSession
      summary: Read the current authenticated owner session
      security:
        - guardianSession: []
      responses:
        '200':
          description: The current server-side session.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          description: The session was denied or expired.
        '426':
          description: TLS is required.
  /v1/auth/logout:
    post:
      operationId: logoutLocalSession
      summary: Revoke and clear the current owner session
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: The current session was revoked and its cookie cleared.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '426':
          description: TLS is required.
  /v1/auth/csrf:
    post:
      operationId: reissueSessionCSRF
      summary: Issue a new synchronizer CSRF proof for the current session
      description: >-
        Change proposal 0003. The console keeps the CSRF proof in memory only,
        so a reload leaves a valid session that cannot mutate anything. This
        returns a replacement without requiring MFA.


        It requires a valid, unexpired, unrevoked session cookie and an exact
        origin match, and nothing else. It carries no CSRF token because
        supplying one is precisely what the caller cannot do; the SameSite
        cookie and the origin check stand in for it. It does not extend the
        absolute session lifetime, is rate limited on the same persistent
        throttle the login path uses, is audited, and returns no credential
        material other than the new proof.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: A new synchronizer proof for the same session.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthCSRFProof'
        '401':
          description: Session or exact-origin validation failed.
        '429':
          description: The authentication throttle is engaged for this account or source.
        '426':
          description: TLS is required.
  /v1/auth/sessions:
    get:
      operationId: listLocalSessions
      summary: List at most 200 owner sessions, including revoked history
      security:
        - guardianSession: []
      responses:
        '200':
          description: Newest-first bounded session history.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionPage'
        '401':
          description: The current session was denied.
        '426':
          description: TLS is required.
  /v1/auth/sessions/{sessionId}:
    delete:
      operationId: revokeLocalSession
      summary: Revoke one session owned by the current local owner
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: The target session was revoked.
        '401':
          description: Session, CSRF, origin, or target ownership validation failed.
        '426':
          description: TLS is required.
  /v1/auth/password:
    post:
      operationId: changeLocalPassword
      summary: Change the owner password and rotate the current session
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthPasswordChangeRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthSessionCredentialsResponse'
        '400':
          description: The replacement password does not meet policy.
        '401':
          description: Current password, session, CSRF, or origin validation failed.
        '426':
          description: TLS is required.
  /v1/organization:
    get:
      operationId: getOrganizationSingleton
      summary: Read the immutable Phase 1 organization singleton
      security:
        - guardianSession: []
      responses:
        '200':
          description: The implicit organization for the authenticated local owner.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '401':
          description: A valid Guardian owner session is required.
        '426':
          description: TLS is required.
  /v1/environments:
    get:
      operationId: listEnvironments
      summary: List environments in the implicit organization
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentListLimit'
      responses:
        '200':
          description: A deterministic bounded environment list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentPage'
        '400':
          description: The list query is invalid.
        '401':
          description: A valid Guardian owner session is required.
        '426':
          description: TLS is required.
    post:
      operationId: createEnvironment
      summary: Create an environment in the implicit organization
      description: >-
        Requires the exact configured HTTPS Origin and synchronizer CSRF token.
        Saving persists and audits configuration only; it never scans or mutates
        the network.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentWriteRequest'
      responses:
        '201':
          $ref: '#/components/responses/EnvironmentResponse'
        '400':
          description: The strict bounded request is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '409':
          description: The normalized environment name already exists.
        '426':
          description: TLS is required.
  /v1/environments/{environmentId}:
    get:
      operationId: getEnvironment
      summary: Read one environment from the implicit organization
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '401':
          description: A valid Guardian owner session is required.
        '404':
          description: The environment does not exist in the implicit organization.
        '426':
          description: TLS is required.
    patch:
      operationId: updateEnvironment
      summary: Rename one environment with optimistic concurrency
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/IfMatchRevision'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentWriteRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '400':
          description: The strict bounded request or ETag is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The environment does not exist in the implicit organization.
        '409':
          description: The normalized environment name already exists.
        '412':
          description: The strong revision ETag is stale.
        '426':
          description: TLS is required.
        '428':
          description: A strong If-Match revision ETag is required.
  /v1/environments/{environmentId}/zones:
    get:
      operationId: listEnvironmentZones
      summary: List private IPv4 zones in one environment
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/EnvironmentListLimit'
      responses:
        '200':
          description: A deterministic bounded zone list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZonePage'
        '400':
          description: The list query is invalid.
        '401':
          description: A valid Guardian owner session is required.
        '404':
          description: The environment does not exist in the implicit organization.
        '426':
          description: TLS is required.
    post:
      operationId: createEnvironmentZone
      summary: Add one canonical RFC1918 IPv4 zone
      description: >-
        Duplicate, nested, or otherwise overlapping prefixes in the same
        environment are rejected. Cross-environment overlap remains valid.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneWriteRequest'
      responses:
        '201':
          $ref: '#/components/responses/ZoneResponse'
        '400':
          description: The name or canonical private IPv4 CIDR is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The environment does not exist in the implicit organization.
        '409':
          description: The normalized name or CIDR conflicts in this environment.
        '426':
          description: TLS is required.
  /v1/environments/{environmentId}/zones/{zoneId}:
    get:
      operationId: getEnvironmentZone
      summary: Read one private IPv4 zone
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/ZoneId'
      responses:
        '200':
          $ref: '#/components/responses/ZoneResponse'
        '401':
          description: A valid Guardian owner session is required.
        '404':
          description: The zone does not exist in the requested environment.
        '426':
          description: TLS is required.
    patch:
      operationId: updateEnvironmentZone
      summary: Update one zone with optimistic concurrency
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/ZoneId'
        - $ref: '#/components/parameters/IfMatchRevision'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneWriteRequest'
      responses:
        '200':
          $ref: '#/components/responses/ZoneResponse'
        '400':
          description: The strict bounded request or ETag is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The zone does not exist in the requested environment.
        '409':
          description: The normalized name or CIDR conflicts in this environment.
        '412':
          description: The strong revision ETag is stale.
        '426':
          description: TLS is required.
        '428':
          description: A strong If-Match revision ETag is required.
    delete:
      operationId: deleteEnvironmentZone
      summary: Remove one zone with optimistic concurrency
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/ZoneId'
        - $ref: '#/components/parameters/IfMatchRevision'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: The zone was removed and the mutation was audited atomically.
        '400':
          description: The ETag is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The zone does not exist in the requested environment.
        '412':
          description: The strong revision ETag is stale.
        '426':
          description: TLS is required.
        '428':
          description: A strong If-Match revision ETag is required.
  /v1/environments/{environmentId}/decoys:
    get:
      operationId: listEnvironmentDecoys
      summary: List decoys in one environment
      description: >-
        Every entry carries the desired object and the observed record as two
        separate fields. They are never merged: a decoy nothing has reported on
        is `unknown`, never `deployed` and never `absent`.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/EnvironmentListLimit'
      responses:
        '200':
          description: A deterministic bounded decoy list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecoyPage'
        '400':
          description: The list query is invalid.
        '401':
          description: A valid Guardian owner session is required.
        '404':
          description: The environment does not exist in the implicit organization.
        '426':
          description: TLS is required.
    post:
      operationId: createEnvironmentDecoy
      summary: Place one decoy in a zone
      description: >-
        The request names a pack by `(pack, pack_version)`. Interaction level
        and digest are resolved from the server-side pack index, and no request
        field accepts a container image reference, command, mount, or
        capability. The address must lie inside the referenced zone's CIDR.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecoyWriteRequest'
      responses:
        '201':
          $ref: '#/components/responses/DecoyResponse'
        '400':
          description: >-
            The request is invalid, the pack version is not in the index, or the
            address is outside the zone.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The environment or zone does not exist.
        '409':
          description: >-
            The normalized name or address conflicts in this environment, or the
            environment already holds the 64 decoys the device channel can carry.
        '426':
          description: TLS is required.
  /v1/environments/{environmentId}/decoys/{decoyId}:
    get:
      operationId: getEnvironmentDecoy
      summary: Read one decoy and its observed record
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/DecoyId'
      responses:
        '200':
          $ref: '#/components/responses/DecoyResponse'
        '401':
          description: A valid Guardian owner session is required.
        '404':
          description: The decoy does not exist in the requested environment.
        '426':
          description: TLS is required.
    patch:
      operationId: updateEnvironmentDecoy
      summary: Update one decoy's configuration with optimistic concurrency
      description: >-
        Configuration only. Enable and disable are separate operations, and
        observed state cannot be written through this or any other REST path.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/DecoyId'
        - $ref: '#/components/parameters/IfMatchRevision'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecoyWriteRequest'
      responses:
        '200':
          $ref: '#/components/responses/DecoyResponse'
        '400':
          description: The strict bounded request, pack reference, placement, or ETag is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The decoy or zone does not exist in the requested environment.
        '409':
          description: The normalized name or address conflicts in this environment.
        '412':
          description: The strong revision ETag is stale.
        '426':
          description: TLS is required.
        '428':
          description: A strong If-Match revision ETag is required.
    delete:
      operationId: deleteEnvironmentDecoy
      summary: Remove one decoy from the active list
      description: >-
        Removal retires the decoy. It is not a deletion of history: the audit
        trail and every interaction already attributed to this decoy identity
        survive it.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/DecoyId'
        - $ref: '#/components/parameters/IfMatchRevision'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: The decoy was removed and the mutation was audited atomically.
        '400':
          description: The ETag is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The decoy does not exist in the requested environment.
        '412':
          description: The strong revision ETag is stale.
        '426':
          description: TLS is required.
        '428':
          description: A strong If-Match revision ETag is required.
  /v1/environments/{environmentId}/decoys/{decoyId}/enable:
    post:
      operationId: enableEnvironmentDecoy
      summary: Ask for this decoy to be deployed
      description: >-
        A separate operation rather than a PATCH field, because WC-D16 assigns
        confirmation levels to operations. Success means the desired state
        changed; it says nothing about what the network is doing.
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/DecoyId'
        - $ref: '#/components/parameters/IfMatchRevision'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          $ref: '#/components/responses/DecoyResponse'
        '400':
          description: The ETag is invalid.
        '401':
          description: Session, CSRF, or exact-origin validation failed.
        '404':
          description: The decoy does not exist in the requested environment.
        '412':
          description: The strong revision ETag is stale.
        '426':
          description: TLS is required.
        '428':
          description: A strong If-Match revision ETag is required.
...
  /v1/environments/{environmentId}/enrollment-tokens/{tokenId}:
    delete:
      operationId: revokeEnrollmentToken
      summary: Revoke one unconsumed enrollment token
      security:
        - guardianSession: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/TokenId'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: The token was atomically revoked.
        '401':
          description: The operator authorization boundary denied the request.
        '409':
          description: The token was not eligible for revocation.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    guardianSession:
      type: apiKey
      in: cookie
      name: __Host-guardian_session
      description: >-
        Opaque same-origin Guardian browser session established by local
        authentication. P1-W2 issues it with Secure, HttpOnly, SameSite=Strict,
        Path=/, and no Domain attribute. The audit boundary accepts at most 512
        bytes before handing the opaque value to its authorizer. Health reads
        are authorized by the same server-side owner session.
    bootstrapToken:
      type: http
      scheme: bearer
      bearerFormat: 32-byte Base64URL one-time token
      description: TLS-only, 15-minute, one-time local-owner bootstrap credential.
    enrollmentToken:
      type: http
      scheme: bearer
      bearerFormat: 32-byte Base64URL one-time token
      description: TLS-only enrollment bootstrap credential; never accepted in a URL.
    deviceMutualTLS:
      type: mutualTLS
      description: Active Guardian product device certificate and durable serial eligibility.
...
  headers:
    NoStore:
      required: true
      schema:
        type: string
        const: no-store
  schemas:
...
    AuthBootstrapRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[a-z0-9._-]+$'
        password:
          type: string
          minLength: 12
          maxLength: 1024
...